moved services, added Casotto logos, askConfirm methods

+++ Casotto-Front-End/src/app/attivita/riga-attivita/riga-attivita.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { RigaCatalogoAttivita } from '../riga-catalogo-attivita';
-import { RigheAttivitaService } from '../righe-attivita.service';
+import { RigaCatalogoAttivita } from 'src/app/model/riga-catalogo-attivita';
+import { RigheAttivitaService } from 'src/app/service/righe-attivita.service';
 
 import { ActivatedRoute, Router } from '@angular/router';
 import { Location } from '@angular/common';
@@ -17,23 +17,24 @@ export class RigaAttiviaComponent implements OnInit {
   //public righe?: Observable<RigaCatalogoAttivita[]>;
 
   constructor(private righeAttivitaService: RigheAttivitaService,
-              private route: ActivatedRoute,  
-              private location: Location,     
-              private router: Router) {
+    private route: ActivatedRoute,
+    private location: Location,
+    private router: Router) {
   }
 
-  public getRigheAttivita():void{ 
-    this.righeAttivitaService.getRighe().subscribe(r => {this.righe = r});
+  public getRigheAttivita(): void {
+    this.righeAttivitaService.getRighe().subscribe(r => { this.righe = r });
     //this.righe =this.righeAttivitaService.getRighe()
   }
 
-  public rigaAttivitaAggiunta(rigaAggiunta : RigaCatalogoAttivita){
+  public rigaAttivitaAggiunta(rigaAggiunta: RigaCatalogoAttivita) {
     this.righe?.push(rigaAggiunta)
     //this.righe?.subscribe(righe => righe.push(rigaAggiunta))//
   }
-  
-  eliminaRiga(riga : RigaCatalogoAttivita){
-    this.righeAttivitaService.removeRiga(riga.id).subscribe(()=> this.getRigheAttivita())
+
+  eliminaRiga(riga: RigaCatalogoAttivita) {
+    if (this.righeAttivitaService.askConfirm("eliminare", "eliminata", "l'", "Attività", "selezionata"))
+      this.righeAttivitaService.removeRiga(riga.id).subscribe(() => this.getRigheAttivita())
     //this.righeAttivitaService.removeRiga(riga.id)
     //  .subscribe(rigaRimossa => {this.righe?.subscribe(righe => righe.filter(r => r ==rigaRimossa as RigaCatalogoAttivita ))})
   }
